+++ src/apps/chifra/internal/scrape/options.go
@@ -26,6 +26,8 @@ type ScrapeOptions struct {
 	Remote     bool                     `json:"remote,omitempty"`     // Pin new chunks to the gateway (requires pinning service keys)
 	Sleep      float64                  `json:"sleep,omitempty"`      // Seconds to sleep between scraper passes
 	StartBlock uint64                   `json:"startBlock,omitempty"` // First block to visit when scraping (snapped back to most recent snap_to_grid mark)
+	RunCount   uint64                   `json:"runCount,omitempty"`   // Run the scraper this many times, then quit
+	DryRun     bool                     `json:"dryRun,omitempty"`     // Show the configuration that would be applied if run,no changes are made
 	Settings   scrapeCfg.ScrapeSettings `json:"settings,omitempty"`   // Configuration items for the scrape
 	Globals    globals.GlobalOptions    `json:"globals,omitempty"`    // The global options
 	Conn       *rpc.Connection          `json:"conn,omitempty"`       // The connection to the RPC server
@@ -45,6 +47,8 @@ func (opts *ScrapeOptions) testLog() {
 	logger.TestLog(opts.Remote, "Remote: ", opts.Remote)
 	logger.TestLog(opts.Sleep != float64(14), "Sleep: ", opts.Sleep)
 	logger.TestLog(opts.StartBlock != 0, "StartBlock: ", opts.StartBlock)
+	logger.TestLog(opts.RunCount != 0, "RunCount: ", opts.RunCount)
+	logger.TestLog(opts.DryRun, "DryRun: ", opts.DryRun)
 	opts.Settings.TestLog(opts.Globals.Chain, opts.Globals.TestMode)
 	opts.Conn.TestLog(opts.getCaches())
 	opts.Globals.TestLog()
@@ -63,6 +67,7 @@ func scrapeFinishParseApi(w http.ResponseWriter, r *http.Request) *ScrapeOptions
 	opts.BlockCnt = 2000
 	opts.Sleep = 14
 	opts.StartBlock = 0
+	opts.RunCount = 0
 	opts.Settings.Apps_per_chunk = 200000
 	opts.Settings.Snap_to_grid = 100000
 	opts.Settings.First_snap = 0
@@ -80,6 +85,10 @@ func scrapeFinishParseApi(w http.ResponseWriter, r *http.Request) *ScrapeOptions
 			opts.Sleep = globals.ToFloat64(value[0])
 		case "startBlock":
 			opts.StartBlock = globals.ToUint64(value[0])
+		case "runCount":
+			opts.RunCount = globals.ToUint64(value[0])
+		case "dryRun":
+			opts.DryRun = true
 		case "appsPerChunk":
 			opts.Settings.Apps_per_chunk = globals.ToUint64(value[0])
 		case "snapToGrid":